possibly done with hatachi multiple sessions. problem with xlsread. wrote method for deleting a subject. unused as of yet

+++ fNIRs/src/fNIRs/Workspace.java
@@ -157,12 +157,31 @@ public class Workspace {
 		preprocess(origFile, freq, hpf, lpf, slideavg, interval);
 		File hbFile = new File("Hb");
 		File hboFile = new File("HbO");
-		File tempFile = addConditions(hbFile, condFile);
-		concatFiles(getHb(name),tempFile);
-		tempFile = addConditions(hboFile, condFile);
-		concatFiles(getHbO(name),tempFile);
-		hbFile.delete();
-		hboFile.delete();
+		File tempFile;
+		if (hbFile.exists()) {
+			tempFile = addConditions(hbFile, condFile);
+			concatFiles(getHb(name),tempFile);
+			hbFile.delete();
+		}
+		if (hboFile.exists()) {
+			tempFile = addConditions(hboFile, condFile);
+			concatFiles(getHbO(name),tempFile);
+			hbFile.delete();
+		}
+	}
+	
+	void concatSession(String name, File hbFile, File hboFile, File condFile) {
+		File tempFile;
+		if (hbFile.exists()) {
+			tempFile = addConditions(hbFile, condFile);
+			concatFiles(getHb(name),tempFile);
+			hbFile.delete();
+		}
+		if (hboFile.exists()) {
+			tempFile = addConditions(hboFile, condFile);
+			concatFiles(getHbO(name),tempFile);
+			hbFile.delete();
+		}
 	}
 	
 	void concatFiles(File file, File other) {
@@ -199,6 +218,25 @@ public class Workspace {
 			return HbOPath;
 		else
 			return null;
+	}
+	
+	boolean deleteDirectory(File dir) {
+		if (dir.isDirectory()) {
+			for (File file : dir.listFiles()) {
+				if (file.isDirectory()) {
+					deleteDirectory(file);
+				}
+				else {
+					file.delete();
+				}
+			}
+		}
+		return dir.delete();
+	}
+	void removeSubject(String name) {
+		String path = subjects.getAbsolutePath();
+		File subject = new File(path + "\\" + name);
+		deleteDirectory(subject);
 	}
 	
 	public static void main(String[] args) {
